+++ src/Pages/Home/Home.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Profile from "../../Components/Home/Profile/Profile";
 import TitleItemBox from "../../Components/Home/TitleItems/TitleItemBox";
-import { setAccessToken } from "../../Api/jwt";
+import { setAccessToken } from "../../Api/axiosInterceptor";
 
 const Home = () => {
   setAccessToken();
